Split amp-youtube component files
